Split Body into partial classes
